+++ kurtosis-module/impl/cl_client_network/cl_client_launcher.go
@@ -1,6 +1,7 @@
 package cl_client_network
 
 import (
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 )
@@ -10,6 +11,7 @@ type ConsensusLayerClientLauncher interface {
 		enclaveCtx *enclaves.EnclaveContext,
 		serviceId services.ServiceID,
 		elClientRpcSockets map[string]bool,  // IP:port of EL client RPC sockets
+		nodeKeystoreDirpaths *prelaunch_data_generator.NodeTypeKeystoreDirpaths,
 	) (
 		resultClientCtx *ConsensusLayerClientContext,
 		resultErr error,
@@ -21,6 +23,7 @@ type ConsensusLayerClientLauncher interface {
 		// NOTE: the ENR of the *consensus layer* boot node
 		bootnodeEnr string,
 		elClientRpcSockets map[string]bool,  // IP:port of EL client RPC sockets
+		nodeKeystoreDirpaths *prelaunch_data_generator.NodeTypeKeystoreDirpaths,
 	) (
 		resultClientCtx *ConsensusLayerClientContext,
 		resultErr error,