+++ kurtosis-module/impl/el_client_network/geth/geth_launcher.go
@@ -32,7 +32,7 @@ const (
 	// NOTE: This can't be 0x00000....000
 	// See: https://github.com/ethereum/go-ethereum/issues/19547
 	miningRewardsAccount = "0x0000000000000000000000000000000000000001"
-	numMiningThreads = 1
+	numMiningThreads = 2
 
 	// The filepath of the genesis JSON file in the shared directory, relative to the shared directory root
 	sharedGenesisJsonRelFilepath = "genesis.json"