+++ kurtosis-module/impl/prelaunch_data_generator/cl_genesis_paths.go
@@ -1,4 +1,4 @@
-package ethereum_genesis_generator
+package prelaunch_data_generator
 
 type CLGenesisPaths struct {
 	// Path to the directory holding all the CL genesis files